Refactor: 프로젝트 구조 리팩토링, README.md 작성
- 프로젝트 구조 정리
- 의존성 문제로 인한 그래프 표시 오류 발생 -> 해결
- README.md 작성

#1
