--- .vscode/settings.json
@@ -1,8 +1,0 @@
-{
-  "github.copilot.enable": {
-    "*": true,
-    "plaintext": false,
-    "markdown": false,
-    "scminput": false
-  }
-}
