+++ requirements.txt
@@ -1,0 +1,6 @@
+streamlit>=1.28.0
+pandas>=2.0.0
+plotly>=5.17.0
+google-api-python-client>=2.100.0
+python-dotenv>=1.0.0
+requests>=2.31.0
