+++ src/__init__.py
@@ -1,0 +1,1 @@
+"""IT 채용정보 분석 프로젝트"""
