+++ src/dashboard/__init__.py
@@ -1,0 +1,1 @@
+"""Streamlit 대시보드 모듈"""
