+++ src/dashboard/search/__init__.py
@@ -1,0 +1,1 @@
+"""외부 검색 API 모듈"""
