+++ src/scrapers/__init__.py
@@ -1,0 +1,1 @@
+"""데이터 스크래핑 모듈"""
